Reliability & Efficiency Changes
Background: Some Zigbee Dimmers (on both Hubitat and SmartThings) will report level changes AFTER the switch is turned off - this will cause extraneous setLevel() commands to be sent to the Slave devices, which will turn the Slave lights back on immediately after they are turned off.

Changes include:
- Allow one device to be designated the master (optional). If set, ONLY changes to this device will be reflected
- Added a separate option to poll the Master device every 5 minutes (not really needed in most cases, as you know).
- Ignore any level changes that occur on the Master when it is turned off
- Eliminate try/catch; instead check whether the device has the desired setLevel() or setSpeed()  commands before trying to invoke the command
- Don't send a setLevel/setSpeed command if the target device is already at the desired level/speed (I probably should have done on/off also).

Note that my code uses the Hubitat-specific form of device.currentValue('attribute', true), which forces the device to return the latest value (which may or may not be the same as the devices' currently cached value). This gives better reliability when multiple events are logged to a device at the same time, as in switch.off/power/level reports that happen within milliseconds on Zigbee dimmers.