Implemented requested changes
This should be pretty close to what you asked for.

The code has to be restructured to enable dynamicPages for the preferences. You may see this as quite massive updates, but in the end it does what you want while dynamically converting existing installations to the new structure without people having to reconfigure.

Let me know if you have any questions or issues!